chore: Seperated Link Logic
